+++ services/download/src/test/java/app/coronawarn/server/services/download/FederationBatchProcessorTest.java
@@ -74,14 +74,6 @@ class FederationBatchProcessorTest {
     batchProcessor = new FederationBatchProcessor(batchInfoService, diagnosisKeyService, federationGatewayClient);
   }
 
-  private BatchDownloadResponse createBatchDownloadResponse(String batchTag, Optional<String> nextBatchTag) {
-    BatchDownloadResponse gatewayResponse = mock(BatchDownloadResponse.class);
-    when(gatewayResponse.getBatchTag()).thenReturn(batchTag);
-    when(gatewayResponse.getNextBatchTag()).thenReturn(nextBatchTag);
-    when(gatewayResponse.getDiagnosisKeyBatch()).thenReturn(FederationBatchUtils.diagnosisKeyBatch);
-    return gatewayResponse;
-  }
-
   @Nested
   @DisplayName("saveFirstBatchInfoForDateTest")
   class SaveFirstBatchInfoForDate {
@@ -95,7 +87,7 @@ class FederationBatchProcessorTest {
 
     @Test
     void testBatchInfoForDateExists() {
-      BatchDownloadResponse serverResponse = createBatchDownloadResponse(batchTag1, Optional.empty());
+      BatchDownloadResponse serverResponse = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.empty());
       when(federationGatewayClient.getDiagnosisKeys(isoDate(date))).thenReturn(serverResponse);
 
       batchProcessor.saveFirstBatchInfoForDate(date);
@@ -120,7 +112,7 @@ class FederationBatchProcessorTest {
     void testOneUnprocessedBatchNoNextBatch() {
       FederationBatchInfo federationBatchInfo = new FederationBatchInfo(batchTag1, date, UNPROCESSED);
       when(batchInfoService.findByStatus(UNPROCESSED)).thenReturn(list(federationBatchInfo));
-      BatchDownloadResponse serverResponse = createBatchDownloadResponse(batchTag1, Optional.empty());
+      BatchDownloadResponse serverResponse = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.empty());
       when(federationGatewayClient.getDiagnosisKeys(batchTag1, isoDate(date))).thenReturn(serverResponse);
 
       batchProcessor.processUnprocessedFederationBatches();
@@ -135,9 +127,9 @@ class FederationBatchProcessorTest {
       FederationBatchInfo batchInfo2 = new FederationBatchInfo(batchTag2, date, UNPROCESSED);
       when(batchInfoService.findByStatus(UNPROCESSED)).thenReturn(list(batchInfo1));
 
-      BatchDownloadResponse serverResponse1 = createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
+      BatchDownloadResponse serverResponse1 = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
       when(federationGatewayClient.getDiagnosisKeys(batchTag1, isoDate(date))).thenReturn(serverResponse1);
-      BatchDownloadResponse serverResponse2 = createBatchDownloadResponse(batchTag2, Optional.empty());
+      BatchDownloadResponse serverResponse2 = FederationBatchUtils.createBatchDownloadResponse(batchTag2, Optional.empty());
       when(federationGatewayClient.getDiagnosisKeys(batchTag2, isoDate(date))).thenReturn(serverResponse2);
 
       batchProcessor.processUnprocessedFederationBatches();
@@ -184,7 +176,7 @@ class FederationBatchProcessorTest {
     @Test
     void testOneErrorBatchNoNextBatch() {
       when(batchInfoService.findByStatus(ERROR)).thenReturn(list(new FederationBatchInfo(batchTag1, date, ERROR)));
-      BatchDownloadResponse serverResponse = createBatchDownloadResponse(batchTag1, Optional.empty());
+      BatchDownloadResponse serverResponse = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.empty());
       when(federationGatewayClient.getDiagnosisKeys(batchTag1, isoDate(date))).thenReturn(serverResponse);
 
       batchProcessor.processErrorFederationBatches();
@@ -202,9 +194,9 @@ class FederationBatchProcessorTest {
 
       when(batchInfoService.findByStatus(ERROR)).thenReturn(list(batchInfo1));
 
-      BatchDownloadResponse serverResponse1 = createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
+      BatchDownloadResponse serverResponse1 = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
       when(federationGatewayClient.getDiagnosisKeys(batchTag1, isoDate(date))).thenReturn(serverResponse1);
-      BatchDownloadResponse serverResponse2 = createBatchDownloadResponse(batchTag2, Optional.empty());
+      BatchDownloadResponse serverResponse2 = FederationBatchUtils.createBatchDownloadResponse(batchTag2, Optional.empty());
       when(federationGatewayClient.getDiagnosisKeys(batchTag2, isoDate(date))).thenReturn(serverResponse2);
 
       batchProcessor.processErrorFederationBatches();
@@ -235,7 +227,7 @@ class FederationBatchProcessorTest {
       when(batchInfoService.findByStatus(ERROR)).thenReturn(list(new FederationBatchInfo(batchTag1, date, ERROR)));
       doThrow(RuntimeException.class).when(batchInfoService).save(any(FederationBatchInfo.class));
 
-      BatchDownloadResponse serverResponse = createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
+      BatchDownloadResponse serverResponse = FederationBatchUtils.createBatchDownloadResponse(batchTag1, Optional.of(batchTag2));
       when(federationGatewayClient.getDiagnosisKeys(batchTag1, isoDate(date))).thenReturn(serverResponse);
 
       batchProcessor.processErrorFederationBatches();
